+++ apps/web/package.json
@@ -15,6 +15,7 @@
     "@stitches/react": "^1.2.8",
     "@vanilla-extract/css": "^1.9.2",
     "apollo-hooks": "*",
+    "classnames": "^2.3.2",
     "deepmerge": "^4.2.2",
     "lodash-es": "^4.17.21",
     "next": "12.3.0",
